Refactor the sql functions into static methods
