+++ libraries/DisplayResults.php
@@ -9,8 +9,9 @@ namespace PMA\libraries;
 
 use PhpMyAdmin\SqlParser\Utils\Query;
 use PMA\libraries\plugins\transformations\Text_Plain_Link;
-use PMA\libraries\URL;
 use PMA\libraries\Sanitize;
+use PMA\libraries\Sql;
+use PMA\libraries\URL;
 
 require_once './libraries/transformations.lib.php';
 
@@ -1429,7 +1430,7 @@ class DisplayResults
             . URL::getHiddenInputs(
                 $this->__get('db'), $this->__get('table')
             )
-            // to avoid calling PMA_handleSortOrder() later
+            // to avoid calling Sql::handleSortOrder() later
             . URL::getHiddenFields(array('sort_by_key' => '1'))
             . __('Sort by key')
             . ': <select name="sql_query" class="autosubmit">' . "\n";
@@ -4256,9 +4257,6 @@ class DisplayResults
         $showtable = $this->__get('showtable');
         $printview = $this->__get('printview');
 
-        // why was this called here? (already called from sql.php)
-        //$this->setConfigParamsForDisplayTable();
-
         /**
          * @todo move this to a central place
          * @todo for other future table types
@@ -4266,9 +4264,7 @@ class DisplayResults
         $is_innodb = (isset($showtable['Type'])
             && $showtable['Type'] == self::TABLE_TYPE_INNO_DB);
 
-        if ($is_innodb
-            && PMA_isJustBrowsing($analyzed_sql_results, true)
-        ) {
+        if ($is_innodb && Sql::isJustBrowsing($analyzed_sql_results, true)) {
             // "j u s t   b r o w s i n g"
             $pre_count = '~';
             $after_count = Util::showHint(
