+++ test/libraries/PMA_mult_submits_test.php
@@ -12,15 +12,10 @@
 use PMA\libraries\Theme;
 use PMA\libraries\URL;
 
-
 require_once 'libraries/mult_submits.lib.php';
-
 require_once 'libraries/database_interface.inc.php';
-
 require_once 'libraries/relation_cleanup.lib.php';
 require_once 'libraries/relation.lib.php';
-require_once 'libraries/sql.lib.php';
-
 
 /**
  * class PMA_MultSubmits_Test
